@@ -1,4 +1,4 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 __title__ = "groq"
-__version__ = "0.31.0"  # x-release-please-version
+__version__ = "0.31.1"  # x-release-please-version
